requirements.txt
configparser==7.0.0
pyaes==1.6.1
pyasn1==0.6.0
PySocks==1.7.1
rsa==4.9
Telethon==1.36.0
toml==0.10.2
